+++ crates/vela_engine/src/tests/source_reload/changed_file_native_method.rs
@@ -190,7 +190,7 @@ fn runtime_stages_changed_file_native_parameter_rejection_until_safe_point() {
         .execution_profile(ExecutionProfile::trusted())
         .register_native_fn(
             NativeFunctionDesc::new("game::native::grant_bonus", NativeFunctionId::new(22))
-                .param("amount", TypeHint::Int),
+                .param("amount", TypeHint::i64()),
             |_| Ok(OwnedValue::Null),
         )
         .build()
@@ -202,7 +202,7 @@ fn runtime_stages_changed_file_native_parameter_rejection_until_safe_point() {
         .execution_profile(ExecutionProfile::trusted())
         .register_native_fn(
             NativeFunctionDesc::new("game::native::grant_bonus", NativeFunctionId::new(22))
-                .param("amount", TypeHint::Float),
+                .param("amount", TypeHint::f64()),
             |_| Ok(OwnedValue::Null),
         )
         .build()
@@ -262,10 +262,10 @@ fn runtime_stages_changed_file_native_parameter_rejection_until_safe_point() {
     assert_eq!(function, "game::native::grant_bonus");
     assert_eq!(old.len(), 1);
     assert_eq!(old[0].name, "amount");
-    assert_eq!(old[0].type_hint.as_deref(), Some("int"));
+    assert_eq!(old[0].type_hint.as_deref(), Some("i64"));
     assert_eq!(new.len(), 1);
     assert_eq!(new[0].name, "amount");
-    assert_eq!(new[0].type_hint.as_deref(), Some("float"));
+    assert_eq!(new[0].type_hint.as_deref(), Some("f64"));
     assert!(source_span.is_none());
     assert_eq!(
         runtime.call_raw(
@@ -299,7 +299,7 @@ fn runtime_stages_changed_file_native_path_proxy_parameter_rejection_until_safe_
         .execution_profile(ExecutionProfile::trusted())
         .register_native_fn(
             NativeFunctionDesc::new("game::native::inspect_path", NativeFunctionId::new(23))
-                .param("path", TypeHint::Int),
+                .param("path", TypeHint::i64()),
             |_| Ok(OwnedValue::Null),
         )
         .build()
@@ -362,7 +362,7 @@ fn runtime_stages_changed_file_native_path_proxy_parameter_rejection_until_safe_
     assert_eq!(old[0].type_hint.as_deref(), Some("path_proxy"));
     assert_eq!(new.len(), 1);
     assert_eq!(new[0].name, "path");
-    assert_eq!(new[0].type_hint.as_deref(), Some("int"));
+    assert_eq!(new[0].type_hint.as_deref(), Some("i64"));
     assert!(source_span.is_none());
     assert_eq!(
         runtime.call_raw(
@@ -384,7 +384,7 @@ fn runtime_stages_changed_file_native_return_rejection_until_safe_point() {
         .execution_profile(ExecutionProfile::trusted())
         .register_native_fn(
             NativeFunctionDesc::new("game::native::grant_bonus", NativeFunctionId::new(22))
-                .returns(TypeHint::Int),
+                .returns(TypeHint::i64()),
             |_| Ok(OwnedValue::Null),
         )
         .build()
@@ -396,7 +396,7 @@ fn runtime_stages_changed_file_native_return_rejection_until_safe_point() {
         .execution_profile(ExecutionProfile::trusted())
         .register_native_fn(
             NativeFunctionDesc::new("game::native::grant_bonus", NativeFunctionId::new(22))
-                .returns(TypeHint::Float),
+                .returns(TypeHint::f64()),
             |_| Ok(OwnedValue::Null),
         )
         .build()
@@ -451,8 +451,8 @@ fn runtime_stages_changed_file_native_return_rejection_until_safe_point() {
         panic!("expected changed native function return ABI");
     };
     assert_eq!(function, "game::native::grant_bonus");
-    assert_eq!(old.as_deref(), Some("int"));
-    assert_eq!(new.as_deref(), Some("float"));
+    assert_eq!(old.as_deref(), Some("i64"));
+    assert_eq!(new.as_deref(), Some("f64"));
     assert!(source_span.is_none());
     assert_eq!(
         runtime.call_raw(
@@ -486,7 +486,7 @@ fn runtime_stages_changed_file_native_path_proxy_return_rejection_until_safe_poi
         .execution_profile(ExecutionProfile::trusted())
         .register_native_fn(
             NativeFunctionDesc::new("game::native::inspect_path", NativeFunctionId::new(23))
-                .returns(TypeHint::Int),
+                .returns(TypeHint::i64()),
             |_| Ok(OwnedValue::Null),
         )
         .build()
@@ -542,7 +542,7 @@ fn runtime_stages_changed_file_native_path_proxy_return_rejection_until_safe_poi
     };
     assert_eq!(function, "game::native::inspect_path");
     assert_eq!(old.as_deref(), Some("path_proxy"));
-    assert_eq!(new.as_deref(), Some("int"));
+    assert_eq!(new.as_deref(), Some("i64"));
     assert!(source_span.is_none());
     assert_eq!(
         runtime.call_raw(
@@ -601,7 +601,7 @@ fn runtime_stages_changed_file_native_stable_id_rename_until_safe_point() {
         .execution_profile(ExecutionProfile::trusted())
         .register_native_fn(
             NativeFunctionDesc::new("game::native::grant_bonus", NativeFunctionId::new(22))
-                .returns(TypeHint::Int)
+                .returns(TypeHint::i64())
                 .effects(EffectSet::host_read()),
             |_| Ok(OwnedValue::Scalar(vela_common::ScalarValue::I64(5))),
         )
@@ -614,7 +614,7 @@ fn runtime_stages_changed_file_native_stable_id_rename_until_safe_point() {
         .execution_profile(ExecutionProfile::trusted())
         .register_native_fn(
             NativeFunctionDesc::new("game::native::grant_bonus_v2", NativeFunctionId::new(22))
-                .returns(TypeHint::Int)
+                .returns(TypeHint::i64())
                 .effects(EffectSet::host_read()),
             |_| Ok(OwnedValue::Scalar(vela_common::ScalarValue::I64(5))),
         )
@@ -732,9 +732,9 @@ fn runtime_stages_changed_file_method_parameter_rejection_until_safe_point() {
     let kind = changed_file_method_rejection_kind(
         "runtime_stage_changed_file_method_parameter",
         MethodDesc::new(HostMethodId::new(9), "grant_exp")
-            .param(MethodParamDesc::new("amount").type_hint("int")),
+            .param(MethodParamDesc::new("amount").type_hint("i64")),
         MethodDesc::new(HostMethodId::new(9), "grant_exp")
-            .param(MethodParamDesc::new("amount").type_hint("float")),
+            .param(MethodParamDesc::new("amount").type_hint("f64")),
         "reload.method.parameter_abi_changed",
     );
 
@@ -752,10 +752,10 @@ fn runtime_stages_changed_file_method_parameter_rejection_until_safe_point() {
     assert_eq!(method, "grant_exp");
     assert_eq!(old.len(), 1);
     assert_eq!(old[0].name, "amount");
-    assert_eq!(old[0].type_hint.as_deref(), Some("int"));
+    assert_eq!(old[0].type_hint.as_deref(), Some("i64"));
     assert_eq!(new.len(), 1);
     assert_eq!(new[0].name, "amount");
-    assert_eq!(new[0].type_hint.as_deref(), Some("float"));
+    assert_eq!(new[0].type_hint.as_deref(), Some("f64"));
     assert!(source_span.is_none());
 }
 
@@ -763,7 +763,7 @@ fn runtime_stages_changed_file_method_parameter_rejection_until_safe_point() {
 fn runtime_stages_changed_file_method_return_rejection_until_safe_point() {
     let kind = changed_file_method_rejection_kind(
         "runtime_stage_changed_file_method_return",
-        MethodDesc::new(HostMethodId::new(9), "grant_exp").return_type("int"),
+        MethodDesc::new(HostMethodId::new(9), "grant_exp").return_type("i64"),
         MethodDesc::new(HostMethodId::new(9), "grant_exp").return_type("null"),
         "reload.method.return_abi_changed",
     );
@@ -780,7 +780,7 @@ fn runtime_stages_changed_file_method_return_rejection_until_safe_point() {
     };
     assert_eq!(type_name, "Player");
     assert_eq!(method, "grant_exp");
-    assert_eq!(old.as_deref(), Some("int"));
+    assert_eq!(old.as_deref(), Some("i64"));
     assert_eq!(new.as_deref(), Some("null"));
     assert!(source_span.is_none());
 }
